+++ EasyGames/Models/ItemDetails.cs
@@ -7,6 +7,6 @@ public class ItemDetails
 
     public List<Review>? Reviews { get; set; }
     public Dictionary<string, string>? Reviewers { get; set; }
-    public int Rating { get; set; } = 0;
+    public double Rating { get; set; } = 0;
     public int RatingCount { get; set; } = 0;
 }
